# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FinIntel Core REST API
  description: Canonical REST API contract between Next.js Web frontend client and Go Core API.
  version: 1.0.0
  license:
    name: Proprietary
    identifier: Proprietary
servers:
  - url: http://localhost:8080
    description: Local Development Server
security:
  - BearerAuth: []
paths:
...
  /api/v1/users/me:
    get:
      summary: Get Current Authenticated User Profile
      description: Retrieves profile details of the authenticated user.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/v1/organizations:
    post:
      summary: Create Organization
      description: Onboards a new organization workspace and assigns creating user as OWNER.
      operationId: createOrganization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization onboarded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
    get:
      summary: List User Organizations
      description: Retrieves all organizations where user is an active member.
      operationId: listOrganizations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user organizations retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationListResponse'
  /api/v1/organizations/{organizationId}:
    get:
      summary: Get Organization Details
      description: Retrieves organization details by ID.
      operationId: getOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Organization details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
  /api/v1/organizations/{organizationId}/accounts:
    get:
      summary: List Chart of Accounts
      description: Retrieves accounts for an organization.
      operationId: listAccounts
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountListResponse'
    post:
      summary: Create Account
      description: Creates a custom account in Chart of Accounts.
      operationId: createAccount
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
...
  /api/v1/organizations/{organizationId}/journal-entries:
    get:
      summary: List Journal Entries
      description: Retrieves general ledger entries for an organization.
      operationId: listJournalEntries
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: List of journal entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryListResponse'
    post:
      summary: Post Journal Entry
      description: Posts a balanced double-entry journal entry enforcing sum(Debits) == sum(Credits).
      operationId: postJournalEntry
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalEntryRequest'
      responses:
        '201':
          description: Journal entry posted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryResponse'
        '400':
          description: Unbalanced entry or invalid parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/organizations/{organizationId}/journal-entries/{entryId}:
    get:
      summary: Get Journal Entry Details
      description: Retrieves details of a specific journal entry.
      operationId: getJournalEntry
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: entryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Journal entry details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryResponse'
  /api/v1/organizations/{organizationId}/staged-transactions:
    get:
      summary: List Staged Transactions
      description: Retrieves staged bank transactions filtered by status (PENDING, APPROVED, REJECTED, POSTED).
      operationId: listStagedTransactions
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ALL, PENDING, APPROVED, REJECTED, POSTED]
      responses:
        '200':
          description: List of staged transactions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StagedTransactionListResponse'
  /api/v1/organizations/{organizationId}/staged-transactions/upload:
    post:
      summary: Upload CSV Bank Statement
      description: Ingests and stages bank CSV transactions with SHA-256 deduplication and auto-categorization.
      operationId: uploadCSV
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: CSV processed and staged successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CSVUploadResponse'
  /api/v1/organizations/{organizationId}/staged-transactions/{id}/approve:
    post:
      summary: Approve Staged Transaction
      description: Approves a staged transaction and assigns target account ID.
      operationId: approveStagedTransaction
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveStagedTransactionRequest'
      responses:
        '200':
          description: Transaction approved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StagedTransactionResponse'
  /api/v1/organizations/{organizationId}/staged-transactions/{id}/reject:
    post:
      summary: Reject Staged Transaction
      description: Rejects a staged transaction.
      operationId: rejectStagedTransaction
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction rejected successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StagedTransactionResponse'
...
  /api/v1/organizations/{organizationId}/reports/trial-balance:
    get:
      summary: Get Trial Balance Report
      description: Generates trial balance aggregating debit and credit totals for all accounts as of target date.
      operationId: getTrialBalance
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: asOfDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Trial balance report generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrialBalanceResponse'
  /api/v1/organizations/{organizationId}/reports/income-statement:
    get:
      summary: Get Income Statement (P&L) Report
      description: Generates income statement summarizing revenues, expenses, and net income within date range.
      operationId: getIncomeStatement
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Income statement report generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomeStatementResponse'
  /api/v1/organizations/{organizationId}/reports/balance-sheet:
    get:
      summary: Get Balance Sheet Report
      description: Generates balance sheet report with Balance Sheet Equation Safeguard (Assets = Liabilities + Equity).
      operationId: getBalanceSheet
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: asOfDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Balance sheet report generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceSheetResponse'
  /api/v1/organizations/{organizationId}/fiscal-periods:
    get:
      summary: List Organization Fiscal Periods
      description: Retrieves list of monthly fiscal periods for an organization.
      operationId: listFiscalPeriods
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of fiscal periods retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiscalPeriodListResponse'
  /api/v1/organizations/{organizationId}/fiscal-periods/generate:
    post:
      summary: Generate Fiscal Periods
      description: Initializes 12 monthly fiscal periods for a target fiscal year.
      operationId: generateFiscalPeriods
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateFiscalPeriodsRequest'
      responses:
        '200':
          description: Fiscal periods generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiscalPeriodListResponse'
...
  /api/v1/organizations/{organizationId}/audit-logs:
    get:
      summary: List Audit Logs
      description: Retrieves atomic audit log entries for organization actions.
      operationId: listAuditLogs
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Audit logs retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
...
  /api/v1/organizations/{organizationId}/anomalies:
    get:
      summary: List Detected Accounting Anomalies
      description: Scans journal entries and staged transactions for duplicates, outliers, and unmapped payees.
      operationId: listAnomalies
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of anomalies retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomalyListResponse'
...
  /api/v1/organizations/{organizationId}/export/ledger:
    get:
      summary: Export General Ledger as CSV
      description: Streams complete General Ledger entries and lines in CSV format.
      operationId: exportGeneralLedgerCSV
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: CSV file generated successfully.
          content:
            text/csv:
              schema:
                type: string
  /api/v1/organizations/{organizationId}/export/audit:
    get:
      summary: Export Audit Trail as CSV
      description: Streams organization audit logs in CSV format.
      operationId: exportAuditLogsCSV
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: CSV file generated successfully.
...
  /api/v1/organizations/{organizationId}/members/{memberId}/role:
    put:
      summary: Update Member Role
      description: Updates the canonical role of an existing organization member.
      operationId: updateMemberRole
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [OWNER, ADMINISTRATOR, ACCOUNTANT, ANALYST, AUDITOR_VIEWER]
      responses:
        '200':
          description: Role updated successfully.
        '400':
          description: Invalid input or last owner demotion prohibited.
        '403':
          description: Permission denied.
  /api/v1/organizations/{organizationId}/members/{memberId}:
    delete:
      summary: Remove Member
      description: Removes a member from an organization.
      operationId: removeMember
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed successfully.
        '400':
          description: Cannot remove last owner.
        '403':
          description: Permission denied.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      required:
        - id
        - email
        - fullName
        - identityProviderIssuer
        - externalSubjectId
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        fullName:
          type: string
        identityProviderIssuer:
          type: string
        externalSubjectId:
          type: string
        createdAt:
          type: string
    CreateOrganizationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        baseCurrency:
          type: string
        fiscalYearStartMonth:
          type: integer
    OrganizationResponse:
      type: object
      required:
        - id
        - name
        - baseCurrency
        - fiscalYearStartMonth
        - role
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        baseCurrency:
          type: string
        fiscalYearStartMonth:
          type: integer
        role:
          type: string
        createdAt:
          type: string
    OrganizationListResponse:
      type: object
      required:
        - organizations
      properties:
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationResponse'
    AccountResponse:
      type: object
      required:
        - id
        - organizationId
        - accountCode
        - name
        - accountType
        - isActive
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        accountCode:
          type: string
        name:
          type: string
        accountType:
          type: string
          enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
        isActive:
          type: boolean
    CreateAccountRequest:
      type: object
      required:
        - accountCode
        - name
        - accountType
      properties:
        accountCode:
          type: string
        name:
          type: string
        accountType:
          type: string
          enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
    AccountListResponse:
      type: object
      required:
        - accounts
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountResponse'
...
    CreateJournalEntryRequest:
      type: object
      required:
        - transactionDate
        - description
        - lines
      properties:
        transactionDate:
          type: string
          format: date
        description:
          type: string
        lines:
          type: array
          minItems: 2
          items:
            $ref: '#/components/schemas/JournalEntryLineRequest'
    JournalEntryResponse:
      type: object
      required:
        - id
        - organizationId
        - entryNumber
        - transactionDate
        - description
        - status
        - createdAt
        - lines
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        entryNumber:
          type: integer
        transactionDate:
          type: string
          format: date
        description:
          type: string
        status:
          type: string
        postedByUserId:
          type: string
        createdAt:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/JournalEntryLineResponse'
    JournalEntryListResponse:
      type: object
      required:
        - entries
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/JournalEntryResponse'
        next_cursor:
          type: [string, "null"]
    StagedTransactionResponse:
      type: object
      required:
        - id
        - organizationId
        - transactionDate
        - description
        - amountMinorUnits
        - rawDataHash
        - status
        - confidenceScore
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        transactionDate:
          type: string
          format: date
        description:
          type: string
        amountMinorUnits:
          type: integer
        rawDataHash:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, POSTED]
        suggestedAccountId:
          type: string
          format: uuid
        suggestedAccountCode:
          type: string
        suggestedAccountName:
          type: string
        confidenceScore:
          type: number
        createdAt:
          type: string
    StagedTransactionListResponse:
      type: object
      required:
        - stagedTransactions
      properties:
        stagedTransactions:
          type: array
          items:
            $ref: '#/components/schemas/StagedTransactionResponse'
    ApproveStagedTransactionRequest:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
    CSVUploadResponse:
      type: object
      required:
        - insertedCount
        - message
      properties:
        insertedCount:
          type: integer
        message:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
...
    TrialBalanceResponse:
      type: object
      required:
        - asOfDate
        - accounts
        - totalDebitsMinorUnits
        - totalCreditsMinorUnits
        - isBalanced
      properties:
        asOfDate:
          type: string
          format: date
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountReportLine'
        totalDebitsMinorUnits:
          type: integer
        totalCreditsMinorUnits:
          type: integer
        isBalanced:
          type: boolean
    IncomeStatementResponse:
      type: object
      required:
        - startDate
        - endDate
        - revenueAccounts
        - expenseAccounts
        - totalRevenueMinorUnits
        - totalExpensesMinorUnits
        - netIncomeMinorUnits
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        revenueAccounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountReportLine'
        expenseAccounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountReportLine'
        totalRevenueMinorUnits:
          type: integer
        totalExpensesMinorUnits:
          type: integer
        netIncomeMinorUnits:
          type: integer
...